# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Couture POS API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "Terminal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "hasOpenShift": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "cashMode": {
            "type": "string",
            "enum": [
              "cash",
              "none"
            ]
          },
          "isPaired": {
            "type": "boolean"
          },
          "isCurrentDevice": {
            "type": "boolean"
          },
          "deviceLastSeenAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "activeCashierName": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "isActive",
          "hasOpenShift",
          "createdAt"
        ]
      },
...
      "CreateTerminalRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "cashMode": {
            "type": "string",
            "enum": [
              "cash",
              "none"
            ],
            "default": "cash"
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
      "UpdateTerminalRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "isActive": {
            "type": "boolean"
          },
          "cashMode": {
            "type": "string",
            "enum": [
              "cash",
              "none"
            ]
          }
        },
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/terminals": {
      "get": {
        "description": "The tenant's counters/tills. Readable by every role so a cashier can pick one when opening a shift.",
        "responses": {
          "200": {
            "description": "Terminals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Terminal"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Create a counter. Names are unique per tenant case-insensitively. Requires manager or owner role.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTerminalRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Terminal created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Terminal"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "A counter with that name already exists"
          }
        }
      }
    },
...
    "/terminals/{terminalId}/pair": {
      "post": {
        "description": "Pair or reassign this browser/device to a counter. Manager/owner only.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "terminalId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Device paired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Terminal"
                }
              }
            }
          },
          "404": {
            "description": "Counter not found"
          },
          "409": {
            "description": "Counter is turned off"
          }
        }
      }
    },
    "/terminals/{terminalId}": {
      "patch": {
        "description": "Rename a counter or turn it on/off. A counter with a shift open on it cannot be turned off. Requires manager or owner role.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "terminalId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTerminalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Terminal updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Terminal"
                }
              }
            }
          },
          "404": {
            "description": "Counter not found"
          },
          "409": {
            "description": "Name already taken, or the counter has a shift open on it"
          }
        }
      },
      "delete": {
        "description": "Delete a counter that has no shift history. Counters with history must be turned off instead so their Z reports keep naming them. Requires manager or owner role.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "terminalId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Terminal deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "deleted"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Counter not found"
          },
          "409": {
            "description": "Counter has shift history and cannot be deleted"
          }
        }
      }
    },
...
  },
...
}